+++ component/src/main/java/io/siddhi/extension/execution/unique/UniqueTimeLengthBatchWindowProcessor.java
@@ -1,12 +1,12 @@
 /*
- * Copyright (c) 2016, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ * Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
  *
  * WSO2 Inc. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
  * in compliance with the License.
  * You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
@@ -15,52 +15,53 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.wso2.extension.siddhi.execution.unique;
+package io.siddhi.extension.execution.unique;
 
-import org.wso2.siddhi.annotation.Example;
-import org.wso2.siddhi.annotation.Extension;
-import org.wso2.siddhi.annotation.Parameter;
-import org.wso2.siddhi.annotation.util.DataType;
-import org.wso2.siddhi.core.config.SiddhiAppContext;
-import org.wso2.siddhi.core.event.ComplexEvent;
-import org.wso2.siddhi.core.event.ComplexEventChunk;
-import org.wso2.siddhi.core.event.state.StateEvent;
-import org.wso2.siddhi.core.event.stream.StreamEvent;
-import org.wso2.siddhi.core.event.stream.StreamEventCloner;
-import org.wso2.siddhi.core.executor.ConstantExpressionExecutor;
-import org.wso2.siddhi.core.executor.ExpressionExecutor;
-import org.wso2.siddhi.core.executor.VariableExpressionExecutor;
-import org.wso2.siddhi.core.query.processor.Processor;
-import org.wso2.siddhi.core.query.processor.SchedulingProcessor;
-import org.wso2.siddhi.core.query.processor.stream.window.FindableProcessor;
-import org.wso2.siddhi.core.query.processor.stream.window.WindowProcessor;
-import org.wso2.siddhi.core.table.Table;
-import org.wso2.siddhi.core.util.Scheduler;
-import org.wso2.siddhi.core.util.collection.operator.CompiledCondition;
-import org.wso2.siddhi.core.util.collection.operator.MatchingMetaInfoHolder;
-import org.wso2.siddhi.core.util.collection.operator.Operator;
-import org.wso2.siddhi.core.util.config.ConfigReader;
-import org.wso2.siddhi.core.util.parser.OperatorParser;
-import org.wso2.siddhi.query.api.definition.Attribute;
-import org.wso2.siddhi.query.api.exception.SiddhiAppValidationException;
-import org.wso2.siddhi.query.api.expression.Expression;
+import io.siddhi.annotation.Example;
+import io.siddhi.annotation.Extension;
+import io.siddhi.annotation.Parameter;
+import io.siddhi.annotation.util.DataType;
+import io.siddhi.core.config.SiddhiAppContext;
+import io.siddhi.core.event.ComplexEvent;
+import io.siddhi.core.event.ComplexEventChunk;
+import io.siddhi.core.event.state.StateEvent;
+import io.siddhi.core.event.stream.StreamEvent;
+import io.siddhi.core.event.stream.StreamEventCloner;
+import io.siddhi.core.executor.ConstantExpressionExecutor;
+import io.siddhi.core.executor.ExpressionExecutor;
+import io.siddhi.core.executor.VariableExpressionExecutor;
+import io.siddhi.core.query.processor.Processor;
+import io.siddhi.core.query.processor.SchedulingProcessor;
+import io.siddhi.core.query.processor.stream.window.FindableProcessor;
+import io.siddhi.core.query.processor.stream.window.WindowProcessor;
+import io.siddhi.core.table.Table;
+import io.siddhi.core.util.Scheduler;
+import io.siddhi.core.util.collection.operator.CompiledCondition;
+import io.siddhi.core.util.collection.operator.MatchingMetaInfoHolder;
+import io.siddhi.core.util.collection.operator.Operator;
+import io.siddhi.core.util.config.ConfigReader;
+import io.siddhi.core.util.parser.OperatorParser;
+import io.siddhi.query.api.definition.Attribute;
+import io.siddhi.query.api.exception.SiddhiAppValidationException;
+import io.siddhi.query.api.expression.Expression;
 
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
 /**
- * class representing unique time batch window processor implementation.
+ * class representing Unique Time Length Batch window processor implementation.
  */
 
 @Extension(
-        name = "timeBatch",
+        name = "timeLengthBatch",
         namespace = "unique",
-        description = "This is a batch (tumbling) time window that is updated "
-                + "with the latest events based on a unique key parameter."
-                + " If a new event that arrives within the time period of a window"
-                + "has a value for the key parameter which matches that of an existing event, "
-                + "the existing event expires and it is replaced by the latest event. ",
+        description = "This is a batch or tumbling time length window that is updated "
+                + "with the latest events based on a unique key parameter. The window tumbles "
+                + "upon the elapse of the time window, or when a number of unique events have arrived."
+                + " If a new event that arrives within the period of the window "
+                + "has a value for the key parameter which matches the value of an existing event, "
+                + "the existing event expires and it is replaced by the new event. ",
         parameters = {
                 @Parameter(name = "unique.key",
                         description = "The attribute that should be checked for uniqueness.",
@@ -68,7 +69,7 @@ import java.util.Map;
                                 DataType.BOOL, DataType.DOUBLE}),
 
                 @Parameter(name = "window.time",
-                        description = "The tumbling time period for which the window should hold events.",
+                        description = "The sliding time period for which the window should hold the events.",
                         type = {DataType.INT, DataType.LONG}),
 
                 @Parameter(name = "start.time",
@@ -76,12 +77,16 @@ import java.util.Map;
                                 " window at a time different to the standard time.",
                         type = {DataType.INT, DataType.LONG},
                         optional = true,
-                        defaultValue = "Timestamp of first event")
+                        defaultValue = "Timestamp of first event"),
+
+                @Parameter(name = "window.length",
+                        description = "The number of events the window should tumble.",
+                        type = {DataType.INT})
         },
         examples = {
                 @Example(
                         syntax = "define stream CseEventStream (symbol string, price float, volume int)\n\n" +
-                                "from CseEventStream#window.unique:timeBatch(symbol, 1 sec)\n" +
+                                "from CseEventStream#window.unique:timeLengthBatch(symbol, 1 sec, 20)\n" +
                                 "select symbol, price, volume\n" +
                                 "insert all events into OutputStream ;",
 
@@ -92,9 +97,11 @@ import java.util.Map;
         }
 )
 
-public class UniqueTimeBatchWindowProcessor extends WindowProcessor implements SchedulingProcessor, FindableProcessor {
+public class UniqueTimeLengthBatchWindowProcessor extends WindowProcessor implements SchedulingProcessor,
+        FindableProcessor {
 
     private long timeInMilliSeconds;
+    private long length;
     private long nextEmitTime = -1;
     private ComplexEventChunk<StreamEvent> currentEventChunk = new ComplexEventChunk<>(false);
     private ComplexEventChunk<StreamEvent> eventsToBeExpired = null;
@@ -105,13 +112,14 @@ public class UniqueTimeBatchWindowProcessor extends WindowProcessor implements S
     private boolean isStartTimeEnabled = false;
     private long startTime = 0;
     private ExpressionExecutor uniqueKeyExpressionExecutor;
+    private boolean eventSent = false;
 
     @Override
     protected void init(ExpressionExecutor[] attributeExpressionExecutors, ConfigReader configReader,
-                        boolean b, SiddhiAppContext siddhiAppContext) {
+                        boolean outputExpectsExpiredEvents, SiddhiAppContext siddhiAppContext) {
         this.siddhiAppContext = siddhiAppContext;
         this.eventsToBeExpired = new ComplexEventChunk<>(false);
-        if (attributeExpressionExecutors.length == 2) {
+        if (attributeExpressionExecutors.length == 3) {
             this.uniqueKeyExpressionExecutor = attributeExpressionExecutors[0];
             if (attributeExpressionExecutors[1] instanceof ConstantExpressionExecutor) {
                 if (attributeExpressionExecutors[1].getReturnType() == Attribute.Type.INT) {
@@ -122,16 +130,33 @@ public class UniqueTimeBatchWindowProcessor extends WindowProcessor implements S
                             .getValue();
                 } else {
                     throw new SiddhiAppValidationException(
-                            "Unique Time Batch window's parameter " + "time should be either"
-                                    + "int or long, but found " + attributeExpressionExecutors[1].getReturnType());
+                            "Unique Time Length Batch window's parameter time should be either int, long or time, " +
+                                    "but found " + attributeExpressionExecutors[1].getReturnType());
                 }
             } else {
-                throw new SiddhiAppValidationException("Unique Time Batch window should have constant "
+                throw new SiddhiAppValidationException("Unique Time Length Batch window should have constant "
                         + "for time parameter but found a dynamic attribute " + attributeExpressionExecutors[1]
                         .getClass().getCanonicalName());
             }
-        } else if (attributeExpressionExecutors.length == 3) {
-                this.uniqueKeyExpressionExecutor = attributeExpressionExecutors[0];
+            if (attributeExpressionExecutors[2] instanceof ConstantExpressionExecutor) {
+                if (attributeExpressionExecutors[2].getReturnType() == Attribute.Type.INT) {
+                    length = (Integer) ((ConstantExpressionExecutor) attributeExpressionExecutors[2])
+                            .getValue();
+                } else if (attributeExpressionExecutors[2].getReturnType() == Attribute.Type.LONG) {
+                    length = (Long) ((ConstantExpressionExecutor) attributeExpressionExecutors[2])
+                            .getValue();
+                } else {
+                    throw new SiddhiAppValidationException(
+                            "Unique Time Length Batch window's parameter " + "length should be either"
+                                    + "int or long, but found " + attributeExpressionExecutors[2].getReturnType());
+                }
+            } else {
+                throw new SiddhiAppValidationException("Unique Time Length Batch window should have constant "
+                        + "for length parameter but found a dynamic attribute " + attributeExpressionExecutors[2]
+                        .getClass().getCanonicalName());
+            }
+        } else if (attributeExpressionExecutors.length == 4) {
+            this.uniqueKeyExpressionExecutor = attributeExpressionExecutors[0];
             if (attributeExpressionExecutors[1] instanceof ConstantExpressionExecutor) {
                 if (attributeExpressionExecutors[1].getReturnType() == Attribute.Type.INT) {
                     timeInMilliSeconds = (Integer) ((ConstantExpressionExecutor) attributeExpressionExecutors[1])
@@ -141,11 +166,12 @@ public class UniqueTimeBatchWindowProcessor extends WindowProcessor implements S
                             .getValue();
                 } else {
                     throw new SiddhiAppValidationException(
-                            "UniqueTimeBatch window's parameter time should be either" + " int or long, but found "
+                            "UniqueTimeLengthBatch window's parameter time should be either" +
+                                    " int or long, but found "
                                     + attributeExpressionExecutors[1].getReturnType());
                 }
             } else {
-                throw new SiddhiAppValidationException("Unique Time Batch window should have constant "
+                throw new SiddhiAppValidationException("Unique Time Length Batch window should have constant "
                         + "for time parameter but found a dynamic attribute " + attributeExpressionExecutors[1]
                         .getClass().getCanonicalName());
             }
@@ -160,18 +186,35 @@ public class UniqueTimeBatchWindowProcessor extends WindowProcessor implements S
                     startTime = Long.parseLong(
                             String.valueOf(((ConstantExpressionExecutor) attributeExpressionExecutors[2]).getValue()));
                 } else {
-                    throw new SiddhiAppValidationException("Expected either "
-                            + "int or long type for UniqueTimeBatch window's start time parameter, but found "
+                    throw new SiddhiAppValidationException("Expected either int or long type for " +
+                            "UniqueTimeLengthBatch window's start time parameter, but found "
                             + attributeExpressionExecutors[2].getReturnType());
                 }
             } else {
-                throw new SiddhiAppValidationException("Unique Time Batch window should have constant "
+                throw new SiddhiAppValidationException("Unique Time Length Batch window should have constant "
                         + "for time parameter but found a dynamic attribute " + attributeExpressionExecutors[2]
                         .getReturnType());
             }
+            if (attributeExpressionExecutors[3] instanceof ConstantExpressionExecutor) {
+                if (attributeExpressionExecutors[3].getReturnType() == Attribute.Type.INT) {
+                    length = (Integer) ((ConstantExpressionExecutor) attributeExpressionExecutors[3])
+                            .getValue();
+                } else if (attributeExpressionExecutors[3].getReturnType() == Attribute.Type.LONG) {
+                    length = (Long) ((ConstantExpressionExecutor) attributeExpressionExecutors[3])
+                            .getValue();
+                } else {
+                    throw new SiddhiAppValidationException(
+                            "Unique Time Length Batch window's parameter " + "length should be either"
+                                    + "int or long, but found " + attributeExpressionExecutors[3].getReturnType());
+                }
+            } else {
+                throw new SiddhiAppValidationException("Unique Time Length Batch window should have constant "
+                        + "for length parameter but found a dynamic attribute " + attributeExpressionExecutors[3]
+                        .getClass().getCanonicalName());
+            }
         } else {
             throw new SiddhiAppValidationException(
-                    "Unique Time Batch window should " + "only have two or three parameters. " + "but found "
+                    "Unique Time Length Batch window should " + "only have three or four parameters. " + "but found "
                             + attributeExpressionExecutors.length + " input attributes");
         }
     }
@@ -191,17 +234,23 @@ public class UniqueTimeBatchWindowProcessor extends WindowProcessor implements S
                     scheduler.notifyAt(nextEmitTime);
                 }
             }
-            boolean sendEvents;
+            boolean sendEventsByTime = false;
+            boolean sendEventsByLength = false;
             if (currentTime >= nextEmitTime) {
                 nextEmitTime += timeInMilliSeconds;
-
                 if (scheduler != null) {
                     scheduler.notifyAt(nextEmitTime);
                 }
-
-                sendEvents = true;
-            } else {
-                sendEvents = false;
+                if (eventSent) { // reset on next batch
+                    eventSent = false;
+                    streamEventChunk.clear();
+                    return;
+                }
+                sendEventsByTime = true;
+            }
+            if (eventSent) { //skip events till next batch
+                streamEventChunk.clear();
+                return;
             }
             while (streamEventChunk.hasNext()) {
                 StreamEvent streamEvent = streamEventChunk.next();
@@ -210,39 +259,17 @@ public class UniqueTimeBatchWindowProcessor extends WindowProcessor implements S
                 }
                 StreamEvent clonedStreamEvent = streamEventCloner.copyStreamEvent(streamEvent);
                 addUniqueEvent(uniqueEventMap, uniqueKeyExpressionExecutor, clonedStreamEvent);
+                if (uniqueEventMap.size() == length) {
+                    sendEventsByLength = true; // emitting batch based on length
+                    break;
+                }
             }
             streamEventChunk.clear();
-            if (sendEvents) {
-                for (StreamEvent event : uniqueEventMap.values()) {
-                    event.setTimestamp(currentTime);
-                    currentEventChunk.add(event);
-                }
-                uniqueEventMap.clear();
-                if (eventsToBeExpired.getFirst() != null) {
-                    while (eventsToBeExpired.hasNext()) {
-                        StreamEvent expiredEvent = eventsToBeExpired.next();
-                        expiredEvent.setTimestamp(currentTime);
-                    }
-                    streamEventChunk.add(eventsToBeExpired.getFirst());
-                }
-                eventsToBeExpired.clear();
-                if (currentEventChunk.getFirst() != null) {
-                    // add reset event in front of current events
-                    streamEventChunk.add(resetEvent);
-                    currentEventChunk.reset();
-                    while (currentEventChunk.hasNext()) {
-                        StreamEvent streamEvent = currentEventChunk.next();
-                        StreamEvent eventClonedForMap = streamEventCloner.copyStreamEvent(streamEvent);
-                        eventClonedForMap.setType(StreamEvent.Type.EXPIRED);
-                        this.eventsToBeExpired.add(eventClonedForMap);
-                    }
-                    if (currentEventChunk.getFirst() != null) {
-                        resetEvent = streamEventCloner.copyStreamEvent(currentEventChunk.getFirst());
-                        resetEvent.setType(ComplexEvent.Type.RESET);
-                        streamEventChunk.add(currentEventChunk.getFirst());
-                    }
-                }
-                currentEventChunk.clear();
+            if (sendEventsByTime || sendEventsByLength) {
+                sendEvents(streamEventChunk, streamEventCloner, currentTime);
+            }
+            if (sendEventsByLength) {
+                eventSent = true; // making events to skip till next time batch
             }
         }
         if (streamEventChunk.getFirst() != null) {
@@ -250,6 +277,40 @@ public class UniqueTimeBatchWindowProcessor extends WindowProcessor implements S
             nextProcessor.process(streamEventChunk);
             streamEventChunk.setBatch(false);
         }
+    }
+
+    private void sendEvents(ComplexEventChunk<StreamEvent> streamEventChunk, StreamEventCloner streamEventCloner,
+                            long currentTime) {
+        for (StreamEvent event : uniqueEventMap.values()) {
+            event.setTimestamp(currentTime);
+            currentEventChunk.add(event);
+        }
+        uniqueEventMap.clear();
+        if (eventsToBeExpired.getFirst() != null) {
+            while (eventsToBeExpired.hasNext()) {
+                StreamEvent expiredEvent = eventsToBeExpired.next();
+                expiredEvent.setTimestamp(currentTime);
+            }
+            streamEventChunk.add(eventsToBeExpired.getFirst());
+        }
+        eventsToBeExpired.clear();
+        if (currentEventChunk.getFirst() != null) {
+            // add reset event in front of current events
+            streamEventChunk.add(resetEvent);
+            currentEventChunk.reset();
+            while (currentEventChunk.hasNext()) {
+                StreamEvent streamEvent = currentEventChunk.next();
+                StreamEvent eventClonedForMap = streamEventCloner.copyStreamEvent(streamEvent);
+                eventClonedForMap.setType(StreamEvent.Type.EXPIRED);
+                this.eventsToBeExpired.add(eventClonedForMap);
+            }
+            if (currentEventChunk.getFirst() != null) {
+                resetEvent = streamEventCloner.copyStreamEvent(currentEventChunk.getFirst());
+                resetEvent.setType(ComplexEvent.Type.RESET);
+                streamEventChunk.add(currentEventChunk.getFirst());
+            }
+        }
+        currentEventChunk.clear();
     }
