+++ component/src/test/java/io/siddhi/extension/execution/unique/UniqueTimeLengthBatchWindowTestCase.java
@@ -1,12 +1,12 @@
 /*
- * Copyright (c) 2016, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ * Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
  *
  * WSO2 Inc. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
  * in compliance with the License.
  * You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
@@ -15,32 +15,32 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.wso2.extension.siddhi.execution.unique;
+package io.siddhi.extension.execution.unique;
 
+import io.siddhi.core.SiddhiAppRuntime;
+import io.siddhi.core.SiddhiManager;
+import io.siddhi.core.event.Event;
+import io.siddhi.core.exception.CannotRestoreSiddhiAppStateException;
+import io.siddhi.core.exception.SiddhiAppCreationException;
+import io.siddhi.core.query.output.callback.QueryCallback;
+import io.siddhi.core.stream.input.InputHandler;
+import io.siddhi.core.util.EventPrinter;
+import io.siddhi.core.util.SiddhiTestHelper;
+import io.siddhi.core.util.persistence.InMemoryPersistenceStore;
+import io.siddhi.core.util.persistence.PersistenceStore;
 import org.apache.log4j.Logger;
 import org.testng.Assert;
 import org.testng.AssertJUnit;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
-import org.wso2.siddhi.core.SiddhiAppRuntime;
-import org.wso2.siddhi.core.SiddhiManager;
-import org.wso2.siddhi.core.event.Event;
-import org.wso2.siddhi.core.exception.CannotRestoreSiddhiAppStateException;
-import org.wso2.siddhi.core.exception.SiddhiAppCreationException;
-import org.wso2.siddhi.core.query.output.callback.QueryCallback;
-import org.wso2.siddhi.core.stream.input.InputHandler;
-import org.wso2.siddhi.core.util.EventPrinter;
-import org.wso2.siddhi.core.util.SiddhiTestHelper;
-import org.wso2.siddhi.core.util.persistence.InMemoryPersistenceStore;
-import org.wso2.siddhi.core.util.persistence.PersistenceStore;
 
 import java.util.concurrent.atomic.AtomicInteger;
 
 /**
- * class representing unique time batch window test case.
+ * class representing unique time length batch window test case.
  */
-public class UniqueTimeBatchWindowTestCase {
-    private static final Logger log = Logger.getLogger(UniqueTimeBatchWindowTestCase.class);
+public class UniqueTimeLengthBatchWindowTestCase {
+    private static final Logger log = Logger.getLogger(UniqueTimeLengthBatchWindowTestCase.class);
     private int inEventCount;
     private int removeEventCount;
     private boolean eventArrived;
@@ -48,7 +48,8 @@ public class UniqueTimeBatchWindowTestCase {
     private int timeout = 30000;
     private AtomicInteger eventCount;
 
-    @BeforeMethod public void init() {
+    @BeforeMethod
+    public void init() {
         inEventCount = 0;
         removeEventCount = 0;
         eventArrived = false;
@@ -56,16 +57,20 @@ public class UniqueTimeBatchWindowTestCase {
 
     }
 
-    @Test public void uniqueTimeWindowBatchTest1() throws InterruptedException {
-        log.info("TimeBatchWindow Test1");
+    @Test
+    public void uniqueTimeLengthWindowBatchTest1() throws InterruptedException {
+        log.info("timeLengthBatchWindow Test1");
         SiddhiManager siddhiManager = new SiddhiManager();
         String cseEventStream = "" + "define stream cseEventStream (symbol string, price float, volume int);";
-        String query = "" + "@info(name = 'query1') " + "from cseEventStream#window.unique:timeBatch(symbol,1 sec) "
-                + "select symbol, price, volume " + "insert all events into outputStream ;";
+        String query = "@info(name = 'query1') " +
+                "from cseEventStream#window.unique:timeLengthBatch(symbol, 1 sec, 5) " +
+                "select symbol, price, volume " +
+                "insert all events into outputStream ;";
 
         SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(cseEventStream + query);
         siddhiAppRuntime.addCallback("query1", new QueryCallback() {
-            @Override public void receive(long timeStamp, Event[] inEvents, Event[] removeEvents) {
+            @Override
+            public void receive(long timeStamp, Event[] inEvents, Event[] removeEvents) {
                 EventPrinter.print(timeStamp, inEvents, removeEvents);
                 if (inEvents != null) {
                     inEventCount = inEventCount + inEvents.length;
@@ -82,14 +87,14 @@ public class UniqueTimeBatchWindowTestCase {
 
         InputHandler inputHandler = siddhiAppRuntime.getInputHandler("cseEventStream");
         siddhiAppRuntime.start();
-        inputHandler.send(new Object[] { "IBM", 700f, 1 });
+        inputHandler.send(new Object[]{"IBM", 700f, 1});
         Thread.sleep(2500);
-        inputHandler.send(new Object[] { "WSO2", 61.5f, 2 });
-        inputHandler.send(new Object[] { "IBM", 700f, 3 });
-        inputHandler.send(new Object[] { "WSO2", 60.5f, 4 });
+        inputHandler.send(new Object[]{"WSO2", 61.5f, 2});
+        inputHandler.send(new Object[]{"IBM", 700f, 3});
+        inputHandler.send(new Object[]{"WSO2", 60.5f, 4});
         Thread.sleep(1100);
-        inputHandler.send(new Object[] { "IBM", 700f, 5 });
-        inputHandler.send(new Object[] { "WSO2", 60.5f, 6 });
+        inputHandler.send(new Object[]{"IBM", 700f, 5});
+        inputHandler.send(new Object[]{"WSO2", 60.5f, 6});
 
         SiddhiTestHelper.waitForEvents(waitTime, 5, eventCount, timeout);
         Assert.assertEquals(inEventCount, 5);
@@ -98,16 +103,69 @@ public class UniqueTimeBatchWindowTestCase {
         siddhiAppRuntime.shutdown();
     }
 
-    @Test public void uniqueTimeWindowBatchTest2() throws InterruptedException {
-        log.info("UniqueTimeBatchWindow Test2");
+    @Test
+    public void uniqueTimeLengthWindowBatchTest15() throws InterruptedException {
+        log.info("timeLengthBatchWindow Test15");
         SiddhiManager siddhiManager = new SiddhiManager();
         String cseEventStream = "" + "define stream cseEventStream (symbol string, price float, volume int);";
-        String query = "" + "@info(name = 'query1') " + "from cseEventStream#window.unique:timeBatch(symbol,1 sec) "
-                + "select symbol, sum(price) as price " + "insert into outputStream ;";
+        String query = "@info(name = 'query1') " +
+                "from cseEventStream#window.unique:timeLengthBatch(symbol, 1 sec, 2) " +
+                "select symbol, price, volume " +
+                "insert all events into outputStream ;";
 
         SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(cseEventStream + query);
         siddhiAppRuntime.addCallback("query1", new QueryCallback() {
-            @Override public void receive(long timeStamp, Event[] inEvents, Event[] removeEvents) {
+            @Override
+            public void receive(long timeStamp, Event[] inEvents, Event[] removeEvents) {
+                EventPrinter.print(timeStamp, inEvents, removeEvents);
+                if (inEvents != null) {
+                    inEventCount = inEventCount + inEvents.length;
+                    eventCount.addAndGet(inEvents.length);
+                }
+                if (removeEvents != null) {
+                    Assert.assertTrue(inEventCount > removeEventCount, "InEvents arrived before RemoveEvents");
+                    removeEventCount = removeEventCount + removeEvents.length;
+                }
+                eventArrived = true;
+            }
+
+        });
+
+        InputHandler inputHandler = siddhiAppRuntime.getInputHandler("cseEventStream");
+        siddhiAppRuntime.start();
+        inputHandler.send(new Object[]{"IBM", 700f, 1});
+        Thread.sleep(2500);
+        inputHandler.send(new Object[]{"WSO2", 61.5f, 2});
+        inputHandler.send(new Object[]{"IBM", 700f, 3});
+        inputHandler.send(new Object[]{"GOOG", 60.5f, 4});
+        Thread.sleep(1100);
+        inputHandler.send(new Object[]{"IBM", 700f, 5});
+        inputHandler.send(new Object[]{"IBM", 705f, 9});
+        inputHandler.send(new Object[]{"WSO2", 60.5f, 6});
+
+        SiddhiTestHelper.waitForEvents(waitTime, 5, eventCount, timeout);
+        Assert.assertEquals(inEventCount, 5);
+        Assert.assertEquals(removeEventCount, 3);
+        Assert.assertTrue(eventArrived);
+        siddhiAppRuntime.shutdown();
+        Thread.sleep(3000);
+    }
+
+    @Test
+    public void uniqueTimeLengthWindowBatchTest2() throws InterruptedException {
+        log.info("UniquetimeLengthBatchWindow Test2");
+        SiddhiManager siddhiManager = new SiddhiManager();
+        String cseEventStream = "" + "define stream cseEventStream (symbol string, price float, volume int);";
+        String query = "" +
+                "@info(name = 'query1') " +
+                "from cseEventStream#window.unique:timeLengthBatch(symbol, 1 sec, 3) " +
+                "select symbol, sum(price) as price " +
+                "insert into outputStream ;";
+
+        SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(cseEventStream + query);
+        siddhiAppRuntime.addCallback("query1", new QueryCallback() {
+            @Override
+            public void receive(long timeStamp, Event[] inEvents, Event[] removeEvents) {
                 EventPrinter.print(timeStamp, inEvents, removeEvents);
 
                 if (inEvents != null) {
@@ -125,14 +183,14 @@ public class UniqueTimeBatchWindowTestCase {
 
         InputHandler inputHandler = siddhiAppRuntime.getInputHandler("cseEventStream");
         siddhiAppRuntime.start();
-        inputHandler.send(new Object[] { "IBM", 700f, 1 });
+        inputHandler.send(new Object[]{"IBM", 700f, 1});
+        Thread.sleep(2100);
+        inputHandler.send(new Object[]{"WSO2", 60.5f, 2});
+        inputHandler.send(new Object[]{"IBM", 700f, 3});
+        inputHandler.send(new Object[]{"WSO2", 60.5f, 4});
         Thread.sleep(1100);
-        inputHandler.send(new Object[] { "WSO2", 60.5f, 2 });
-        inputHandler.send(new Object[] { "IBM", 700f, 3 });
-        inputHandler.send(new Object[] { "WSO2", 60.5f, 4 });
-        Thread.sleep(1100);
-        inputHandler.send(new Object[] { "IBM", 700f, 5 });
-        inputHandler.send(new Object[] { "WSO2", 60.5f, 6 });
+        inputHandler.send(new Object[]{"IBM", 700f, 5});
+        inputHandler.send(new Object[]{"WSO2", 60.5f, 6});
 
         SiddhiTestHelper.waitForEvents(waitTime, 3, eventCount, timeout);
         Assert.assertEquals(inEventCount, 3);
@@ -141,16 +199,21 @@ public class UniqueTimeBatchWindowTestCase {
         siddhiAppRuntime.shutdown();
     }
 
-    @Test public void uniqueTimeWindowBatchTest3() throws InterruptedException {
-        log.info("UniqueTimeBatchWindow Test3");
+    @Test
+    public void uniqueTimeLengthWindowBatchTest3() throws InterruptedException {
+        log.info("UniquetimeLengthBatchWindow Test3");
         SiddhiManager siddhiManager = new SiddhiManager();
         String cseEventStream = "" + "define stream cseEventStream (symbol string, price float, volume int);";
-        String query = "" + "@info(name = 'query1') " + "from cseEventStream#window.unique:timeBatch(symbol,1 sec) "
-                + "select symbol, sum(price) as price,volume " + "insert expired events into outputStream ;";
+        String query = "" +
+                "@info(name = 'query1') " +
+                "from cseEventStream#window.unique:timeLengthBatch(symbol,1 sec, 2) " +
+                "select symbol, sum(price) as price,volume " +
+                "insert expired events into outputStream ;";
 
         SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(cseEventStream + query);
         siddhiAppRuntime.addCallback("query1", new QueryCallback() {
-            @Override public void receive(long timeStamp, Event[] inEvents, Event[] removeEvents) {
+            @Override
+            public void receive(long timeStamp, Event[] inEvents, Event[] removeEvents) {
                 EventPrinter.print(timeStamp, inEvents, removeEvents);
 
                 if (inEvents != null) {
@@ -168,14 +231,14 @@ public class UniqueTimeBatchWindowTestCase {
 
         InputHandler inputHandler = siddhiAppRuntime.getInputHandler("cseEventStream");
         siddhiAppRuntime.start();
-        inputHandler.send(new Object[] { "IBM", 700f, 1 });
+        inputHandler.send(new Object[]{"IBM", 700f, 1});
         Thread.sleep(1100);
-        inputHandler.send(new Object[] { "WSO2", 60.5f, 2 });
-        inputHandler.send(new Object[] { "IBM", 700f, 3 });
-        inputHandler.send(new Object[] { "WSO2", 61.5f, 4 });
+        inputHandler.send(new Object[]{"WSO2", 60.5f, 2});
+        inputHandler.send(new Object[]{"IBM", 700f, 3});
+        inputHandler.send(new Object[]{"WSO2", 61.5f, 4});
         Thread.sleep(1100);
-        inputHandler.send(new Object[] { "IBM", 700f, 5 });
-        inputHandler.send(new Object[] { "WSO2", 60.5f, 6 });
+        inputHandler.send(new Object[]{"IBM", 700f, 5});
+        inputHandler.send(new Object[]{"WSO2", 60.5f, 6});
 
         SiddhiTestHelper.waitForEvents(waitTime, 3, eventCount, timeout);
         Assert.assertEquals(inEventCount, 0);
@@ -184,13 +247,15 @@ public class UniqueTimeBatchWindowTestCase {
         siddhiAppRuntime.shutdown();
     }
 
-    @Test public void uniqueTimeWindowBatchTest4() throws InterruptedException {
-        log.info("UniqueTimeBatchWindow Test4");
+    @Test
+    public void uniqueTimeLengthWindowBatchTest4() throws InterruptedException {
+        log.info("UniquetimeLengthBatchWindow Test4");
         SiddhiManager siddhiManager = new SiddhiManager();
         String streams = "" + "define stream cseEventStream (symbol string, price float, volume int); "
                 + "define stream twitterStream (user string, tweet string, company string); ";
-        String query = "" + "@info(name = 'query1') " + "from cseEventStream#window.unique:timeBatch(symbol,1 sec)"
-                + " join twitterStream#window.unique:timeBatch(company,1 sec) "
+        String query = "" + "@info(name = 'query1') " +
+                "from cseEventStream#window.unique:timeLengthBatch(symbol,1 sec, 3)"
+                + " join twitterStream#window.unique:timeLengthBatch(company,1 sec, 3) "
                 + "on cseEventStream.symbol== twitterStream.company "
                 + "select cseEventStream.symbol as symbol, twitterStream.tweet, cseEventStream.price "
                 + "insert all events into outputStream ;";
@@ -198,7 +263,8 @@ public class UniqueTimeBatchWindowTestCase {
         SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
         try {
             siddhiAppRuntime.addCallback("query1", new QueryCallback() {
-                @Override public void receive(long timeStamp, Event[] inEvents, Event[] removeEvents) {
+                @Override
+                public void receive(long timeStamp, Event[] inEvents, Event[] removeEvents) {
                     EventPrinter.print(timeStamp, inEvents, removeEvents);
 
                     if (inEvents != null) {
@@ -214,11 +280,11 @@ public class UniqueTimeBatchWindowTestCase {
             InputHandler cseEventStreamHandler = siddhiAppRuntime.getInputHandler("cseEventStream");
             InputHandler twitterStreamHandler = siddhiAppRuntime.getInputHandler("twitterStream");
             siddhiAppRuntime.start();
-            cseEventStreamHandler.send(new Object[] { "WSO2", 55.6f, 100 });
-            twitterStreamHandler.send(new Object[] { "User1", "Hello World", "WSO2" });
-            cseEventStreamHandler.send(new Object[] { "IBM", 75.6f, 100 });
+            cseEventStreamHandler.send(new Object[]{"WSO2", 55.6f, 100});
+            twitterStreamHandler.send(new Object[]{"User1", "Hello World", "WSO2"});
+            cseEventStreamHandler.send(new Object[]{"IBM", 75.6f, 100});
             Thread.sleep(1100);
-            cseEventStreamHandler.send(new Object[] { "WSO2", 57.6f, 100 });
+            cseEventStreamHandler.send(new Object[]{"WSO2", 57.6f, 100});
 
             SiddhiTestHelper.waitForEvents(waitTime, 2, eventCount, timeout);
             Assert.assertTrue(inEventCount == 1 || inEventCount == 2, "In Events can be 1 or 2 ");
@@ -229,13 +295,15 @@ public class UniqueTimeBatchWindowTestCase {
         }
     }
 
-    @Test public void uniqueTimeWindowBatchTest5() throws InterruptedException {
-        log.info("UniqueTimeBatchWindow Test5");
+    @Test
+    public void uniqueTimeLengthWindowBatchTest5() throws InterruptedException {
+        log.info("UniquetimeLengthBatchWindow Test5");
         SiddhiManager siddhiManager = new SiddhiManager();
         String streams = "" + "define stream cseEventStream (symbol string, price float, volume int); "
                 + "define stream twitterStream (user string, tweet string, company string); ";
-        String query = "" + "@info(name = 'query1') " + "from cseEventStream#window.unique:timeBatch(symbol,1 sec) "
-                + "join twitterStream#window.unique:timeBatch(company,1 sec) "
+        String query = "" + "@info(name = 'query1') " +
+                "from cseEventStream#window.unique:timeLengthBatch(symbol,1 sec, 3) "
+                + "join twitterStream#window.unique:timeLengthBatch(company,1 sec, 3) "
                 + "on cseEventStream.symbol== twitterStream.company "
                 + "select cseEventStream.symbol as symbol, twitterStream.tweet, cseEventStream.price "
                 + "insert into outputStream ;";
@@ -243,7 +311,8 @@ public class UniqueTimeBatchWindowTestCase {
         SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
         try {
             siddhiAppRuntime.addCallback("query1", new QueryCallback() {
-                @Override public void receive(long timeStamp, Event[] inEvents, Event[] removeEvents) {
+                @Override
+                public void receive(long timeStamp, Event[] inEvents, Event[] removeEvents) {
                     EventPrinter.print(timeStamp, inEvents, removeEvents);
 
                     if (inEvents != null) {
@@ -259,11 +328,11 @@ public class UniqueTimeBatchWindowTestCase {
             InputHandler cseEventStreamHandler = siddhiAppRuntime.getInputHandler("cseEventStream");
             InputHandler twitterStreamHandler = siddhiAppRuntime.getInputHandler("twitterStream");
             siddhiAppRuntime.start();
-            cseEventStreamHandler.send(new Object[] { "WSO2", 55.6f, 100 });
-            twitterStreamHandler.send(new Object[] { "User1", "Hello World", "WSO2" });
-            cseEventStreamHandler.send(new Object[] { "IBM", 75.6f, 100 });
+            cseEventStreamHandler.send(new Object[]{"WSO2", 55.6f, 100});
+            twitterStreamHandler.send(new Object[]{"User1", "Hello World", "WSO2"});
+            cseEventStreamHandler.send(new Object[]{"IBM", 75.6f, 100});
             Thread.sleep(1500);
-            cseEventStreamHandler.send(new Object[] { "WSO2", 57.6f, 100 });
+            cseEventStreamHandler.send(new Object[]{"WSO2", 57.6f, 100});
 
             SiddhiTestHelper.waitForEvents(waitTime, 2, eventCount, timeout);
             Assert.assertTrue(inEventCount == 1 || inEventCount == 2, "In Events can be 1 or 2 ");
@@ -274,17 +343,20 @@ public class UniqueTimeBatchWindowTestCase {
         }
     }
 
-    @Test public void uniqueTimeWindowBatchTest6() throws InterruptedException {
-        log.info("UniqueTimeBatchWindow Test6");
+    @Test
+    public void uniqueTimeLengthWindowBatchTest6() throws InterruptedException {
+        log.info("UniquetimeLengthBatchWindow Test6");
         SiddhiManager siddhiManager = new SiddhiManager();
         String cseEventStream = "" + "define stream cseEventStream (symbol string, price float, volume int);";
-        String query = "" + "@info(name = 'query1') " + "from cseEventStream#window.unique:timeBatch(symbol,2 sec) "
+        String query = "" + "@info(name = 'query1') " +
+                "from cseEventStream#window.unique:timeLengthBatch(symbol,2 sec, 2) "
                 + "select symbol, sum(price) as sumPrice, volume " + "insert into outputStream ;";
 
         SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(cseEventStream + query);
 
         siddhiAppRuntime.addCallback("query1", new QueryCallback() {
-            @Override public void receive(long timeStamp, Event[] inEvents, Event[] removeEvents) {
+            @Override
+            public void receive(long timeStamp, Event[] inEvents, Event[] removeEvents) {
                 EventPrinter.print(timeStamp, inEvents, removeEvents);
 
                 if (inEventCount == 0) {
@@ -308,14 +380,14 @@ public class UniqueTimeBatchWindowTestCase {
         while (System.currentTimeMillis() % 2000 != 0) {
 
         }
-        inputHandler.send(new Object[] { "IBM", 700f, 0 });
-        inputHandler.send(new Object[] { "WSO2", 60.5f, 1 });
+        inputHandler.send(new Object[]{"IBM", 700f, 0});
+        inputHandler.send(new Object[]{"WSO2", 60.5f, 1});
         Thread.sleep(8100);
-        inputHandler.send(new Object[] { "WSO2", 60.5f, 1 });
-        inputHandler.send(new Object[] { "II", 60.5f, 1 });
+        inputHandler.send(new Object[]{"WSO2", 60.5f, 1});
+        inputHandler.send(new Object[]{"II", 60.5f, 1});
         Thread.sleep(13100);
-        inputHandler.send(new Object[] { "TT", 60.5f, 1 });
-        inputHandler.send(new Object[] { "YY", 60.5f, 1 });
+        inputHandler.send(new Object[]{"TT", 60.5f, 1});
+        inputHandler.send(new Object[]{"YY", 60.5f, 1});
 
         SiddhiTestHelper.waitForEvents(waitTime, 3, eventCount, timeout);
         Assert.assertEquals(inEventCount, 3);
@@ -325,22 +397,12 @@ public class UniqueTimeBatchWindowTestCase {
     }
 
     @Test(expectedExceptions = SiddhiAppCreationException.class)
-    public void uniqueTimeWindowBatchTest8() {
-        log.info("TimeBatchWindow Test for Unique Time Batch window invalid type");
+    public void uniqueTimeLengthWindowBatchTest7() {
+        log.info("timeLengthBatchWindow Test for Unique Length Batch window should variable case ");
         SiddhiManager siddhiManager = new SiddhiManager();
         String cseEventStream = "" + "define stream cseEventStream (symbol string, price float, volume int);";
-        String query = "" + "@info(name = 'query1') " + "from cseEventStream#window.unique:timeBatch(symbol,'1 sec') "
-                + "select symbol, price, volume " + "insert all events into outputStream ;";
-
-        siddhiManager.createSiddhiAppRuntime(cseEventStream + query);
-
-    }
-    @Test(expectedExceptions = SiddhiAppCreationException.class)
-    public void uniqueTimeWindowBatchTest9() {
-        log.info("TimeBatchWindow Test for Unique Time Batch window Constant case ");
-        SiddhiManager siddhiManager = new SiddhiManager();
-        String cseEventStream = "" + "define stream cseEventStream (symbol string, price float, volume int);";
-        String query = "" + "@info(name = 'query1') " + "from cseEventStream#window.unique:timeBatch(symbol,volume) "
+        String query = "" + "@info(name = 'query1') " +
+                "from cseEventStream#window.unique:timeLengthBatch('symbol',1 sec, 2) "
                 + "select symbol, price, volume " + "insert all events into outputStream ;";
 
         siddhiManager.createSiddhiAppRuntime(cseEventStream + query);
@@ -348,11 +410,12 @@ public class UniqueTimeBatchWindowTestCase {
     }
 
     @Test(expectedExceptions = SiddhiAppCreationException.class)
-    public void uniqueTimeWindowBatchTest10() {
-        log.info("TimeBatchWindow Test for invalid number of parameters");
+    public void uniqueTimeLengthWindowBatchTest8() {
+        log.info("timeLengthBatchWindow Test for Unique Time Batch window invalid type");
         SiddhiManager siddhiManager = new SiddhiManager();
         String cseEventStream = "" + "define stream cseEventStream (symbol string, price float, volume int);";
-        String query = "" + "@info(name = 'query1') " + "from cseEventStream#window.unique:timeBatch(symbol) "
+        String query = "" + "@info(name = 'query1') " +
+                "from cseEventStream#window.unique:timeLengthBatch(symbol,'1 sec', 2) "
                 + "select symbol, price, volume " + "insert all events into outputStream ;";
 
         siddhiManager.createSiddhiAppRuntime(cseEventStream + query);
@@ -360,12 +423,39 @@ public class UniqueTimeBatchWindowTestCase {
     }
 
     @Test(expectedExceptions = SiddhiAppCreationException.class)
-    public void uniqueTimeWindowBatchTest11() {
-        log.info("TimeBatchWindow Test for Unique Length Batch window variable case length three");
+    public void uniqueTimeLengthWindowBatchTest9() {
+        log.info("timeLengthBatchWindow Test for Unique Time Batch window Constant case ");
+        SiddhiManager siddhiManager = new SiddhiManager();
+        String cseEventStream = "" + "define stream cseEventStream (symbol string, price float, volume int);";
+        String query = "" + "@info(name = 'query1') " +
+                "from cseEventStream#window.unique:timeLengthBatch(symbol,volume) "
+                + "select symbol, price, volume " + "insert all events into outputStream ;";
+
+        siddhiManager.createSiddhiAppRuntime(cseEventStream + query);
+
+    }
+
+    @Test(expectedExceptions = SiddhiAppCreationException.class)
+    public void uniqueTimeLengthWindowBatchTest10() {
+        log.info("timeLengthBatchWindow Test for invalid number of parameters");
+        SiddhiManager siddhiManager = new SiddhiManager();
+        String cseEventStream = "" + "define stream cseEventStream (symbol string, price float, volume int);";
+        String query = "" + "@info(name = 'query1') " +
+                "from cseEventStream#window.unique:timeLengthBatch(symbol) "
+                + "select symbol, price, volume " + "insert all events into outputStream ;";
+
+        siddhiManager.createSiddhiAppRuntime(cseEventStream + query);
+
+    }
+
+    @Test(expectedExceptions = SiddhiAppCreationException.class)
+    public void uniqueTimeLengthWindowBatchTest11() {
+        log.info("timeLengthBatchWindow Test for Unique Length Batch window variable case length three");
         SiddhiManager siddhiManager = new SiddhiManager();
         String cseEventStream = "" + "define stream cseEventStream (symbol string, price float, volume int);";
         String query =
-                "" + "@info(name = 'query1') " + "from cseEventStream#window.unique:timeBatch('symbol',1 sec,volume) "
+                "" + "@info(name = 'query1') " +
+                        "from cseEventStream#window.unique:timeLengthBatch('symbol',1 sec,volume) "
                         + "select symbol, price, volume " + "insert all events into outputStream ;";
 
         siddhiManager.createSiddhiAppRuntime(cseEventStream + query);
@@ -373,12 +463,13 @@ public class UniqueTimeBatchWindowTestCase {
     }
 
     @Test(expectedExceptions = SiddhiAppCreationException.class)
-    public void uniqueTimeWindowBatchTest12() {
-        log.info("TimeBatchWindow Test for UniqueTimeBatch window invalid type case length three");
+    public void uniqueTimeLengthWindowBatchTest12() {
+        log.info("timeLengthBatchWindow Test for UniquetimeLengthBatch window invalid type case length three");
         SiddhiManager siddhiManager = new SiddhiManager();
         String cseEventStream = "" + "define stream cseEventStream (symbol string, price float, volume int);";
         String query =
-                "" + "@info(name = 'query1') " + "from cseEventStream#window.unique:timeBatch(symbol,'1 sec',volume) "
+                "" + "@info(name = 'query1') " +
+                        "from cseEventStream#window.unique:timeLengthBatch(symbol,'1 sec',volume, 5) "
                         + "select symbol, price, volume " + "insert all events into outputStream ;";
 
         siddhiManager.createSiddhiAppRuntime(cseEventStream + query);
@@ -386,30 +477,48 @@ public class UniqueTimeBatchWindowTestCase {
     }
 
     @Test(expectedExceptions = SiddhiAppCreationException.class)
-    public void uniqueTimeWindowBatchTest13() {
-        log.info("TimeBatchWindow Test for Unique Time Batch window should constant case length three");
+    public void uniqueTimeLengthWindowBatchTest13() {
+        log.info("timeLengthBatchWindow Test for Unique Time Batch window should constant case length three");
         SiddhiManager siddhiManager = new SiddhiManager();
         String cseEventStream = "" + "define stream cseEventStream (symbol string, price float, volume int);";
         String query =
-                "" + "@info(name = 'query1') " + "from cseEventStream#window.unique:timeBatch(symbol,price,volume) "
+                "" + "@info(name = 'query1') " +
+                        "from cseEventStream#window.unique:timeLengthBatch(symbol,price,volume, 6) "
                         + "select symbol, price, volume " + "insert all events into outputStream ;";
 
         siddhiManager.createSiddhiAppRuntime(cseEventStream + query);
 
     }
 
-    @Test public void uniqueTimeWindowBatchTest14() throws InterruptedException {
-        log.info("TimeBatchWindow Test1");
+    @Test(expectedExceptions = SiddhiAppCreationException.class)
+    public void uniqueTimeLengthWindowBatchTest16() {
+        log.info("timeLengthBatchWindow Test for Unique Time Batch window should constant case length four");
+        SiddhiManager siddhiManager = new SiddhiManager();
+        String cseEventStream = "" + "define stream cseEventStream (symbol string, price float, volume int);";
+        String query =
+                "" + "@info(name = 'query1') " +
+                        "from cseEventStream#window.unique:timeLengthBatch(symbol,1 sec, 0, volume) "
+                        + "select symbol, price, volume " + "insert all events into outputStream ;";
+
+        siddhiManager.createSiddhiAppRuntime(cseEventStream + query);
+
+    }
+
+    @Test
+    public void uniqueTimeLengthWindowBatchTest14() throws InterruptedException {
+        log.info("timeLengthBatchWindow Test1");
         PersistenceStore persistenceStore = new InMemoryPersistenceStore();
         SiddhiManager siddhiManager = new SiddhiManager();
         siddhiManager.setPersistenceStore(persistenceStore);
         String cseEventStream = "" + "define stream cseEventStream (symbol string, price float, volume int);";
-        String query = "" + "@info(name = 'query1') " + "from cseEventStream#window.unique:timeBatch(symbol,1 sec) "
+        String query = "" + "@info(name = 'query1') " +
+                "from cseEventStream#window.unique:timeLengthBatch(symbol,1 sec, 5) "
                 + "select symbol, price, volume " + "insert all events into outputStream ;";
 
         SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(cseEventStream + query);
         siddhiAppRuntime.addCallback("query1", new QueryCallback() {
-            @Override public void receive(long timeStamp, Event[] inEvents, Event[] removeEvents) {
+            @Override
+            public void receive(long timeStamp, Event[] inEvents, Event[] removeEvents) {
                 EventPrinter.print(timeStamp, inEvents, removeEvents);
                 if (inEvents != null) {
                     inEventCount = inEventCount + inEvents.length;
@@ -426,14 +535,14 @@ public class UniqueTimeBatchWindowTestCase {
 
         InputHandler inputHandler = siddhiAppRuntime.getInputHandler("cseEventStream");
         siddhiAppRuntime.start();
-        inputHandler.send(new Object[] { "IBM", 700f, 1 });
+        inputHandler.send(new Object[]{"IBM", 700f, 1});
         SiddhiTestHelper.waitForEvents(2500, 1, eventCount, timeout);
-        inputHandler.send(new Object[] { "WSO2", 61.5f, 2 });
-        inputHandler.send(new Object[] { "IBM", 700f, 3 });
-        inputHandler.send(new Object[] { "WSO2", 60.5f, 4 });
+        inputHandler.send(new Object[]{"WSO2", 61.5f, 2});
+        inputHandler.send(new Object[]{"IBM", 700f, 3});
+        inputHandler.send(new Object[]{"WSO2", 60.5f, 4});
         SiddhiTestHelper.waitForEvents(1100, 3, eventCount, timeout);
-        inputHandler.send(new Object[] { "IBM", 700f, 5 });
-        inputHandler.send(new Object[] { "WSO2", 60.5f, 6 });
+        inputHandler.send(new Object[]{"IBM", 700f, 5});
+        inputHandler.send(new Object[]{"WSO2", 60.5f, 6});
         SiddhiTestHelper.waitForEvents(waitTime, 6, eventCount, timeout);
         Assert.assertEquals(inEventCount, 5);
         siddhiAppRuntime.persist();
@@ -445,11 +554,11 @@ public class UniqueTimeBatchWindowTestCase {
         } catch (CannotRestoreSiddhiAppStateException e) {
             Assert.fail("Error in restoring last revision");
         }
-        inputHandler.send(new Object[] { "IBM", 700f, 1 });
+        inputHandler.send(new Object[]{"IBM", 700f, 1});
         SiddhiTestHelper.waitForEvents(3500, 6, eventCount, timeout);
-        inputHandler.send(new Object[] { "WSO2", 61.5f, 2 });
-        inputHandler.send(new Object[] { "IBM", 700f, 3 });
-        inputHandler.send(new Object[] { "WSO2", 60.5f, 4 });
+        inputHandler.send(new Object[]{"WSO2", 61.5f, 2});
+        inputHandler.send(new Object[]{"IBM", 700f, 3});
+        inputHandler.send(new Object[]{"WSO2", 60.5f, 4});
         SiddhiTestHelper.waitForEvents(waitTime, 8, eventCount, timeout);
         Assert.assertEquals(inEventCount, 8);
         Assert.assertEquals(removeEventCount, 7);
